Split WeightLiftingMetrics.jsx
All metric components are now split!
